Fix attach node issue, improve logging
Fixing #6, if a subtype has no attach node then just copy the part
prefab's.  Testing whether to actually switch is done by looking at
attach node IDs.  The cause of the bug was that re-rendering the drag
cubes was causing the attach node to become non-null (maybe due to
serialization?).