Change the way current subtype is tracked
There is now only one piece of state on the module that keeps track of
the subtype (the index), and one value that is saved to persist it (the
name).  The field persisted in the config node now has a different value
(currentSubtypeName -> currentSubtype).  Loading an old craft with
either the index or currentSubtypeName defined will still work for now