Fix textures being reset when drag cubes rendered
Textures are reset to prep for copy, but nothing notifies them to reset
(since it relied on KSP calling OnCopy).  Fix is to keep track of the
"original" object in a ScriptableObject (since that will still point to
the original instance once copied) and notify it in OnAwake from the
copy to reset itself.