Deactivate texture switches before re-initializing
If a part up the hierarchy gets its drag cubes rendered when loading a
craft, the texture switches will be activated by the process of creating
the drag cube clone (deactivated before the clone and then activated
after), then Start is called on this part which re-initializes the
texture switches.  The original information is lost and the texture
switch is now locked in.