Force fire part copying methods on child parts
If you copy a part in the editor, these methods aren't fired for child parts, causing modified textures to become permanent.  This fires them so that texture switches properly get disabled before switching and properly re-enabled after.  Hopefully this will be fixed in KSP at some point so that this code can be removed.